Choose File Button rework
Previously, the text showing the currently selected file (or "no file selected") was made to be transparent. That caused the area of the button to extend through the invisible text, and if the user hovered over the text area, a tooltip would appear. Now, the button is hidden, and another visible button triggers the input button to fire intead.