Fix MyEditor.getVerse() when the current selection does not match the desired verse.
When a verse number is clicked on the left side of the number, the selection will actually be at the end of the previous verse's text. Thus, when the user clicks a verse number, the verse number menu may be out of sync with the current selection if the user clicked on the far left. This results in a getVerse() function that is more complicated than desired, because it may need to use both the Editor.above() and Editor.next() functions to find the desired verse. Furthermore, the match function does not work as anticipated when using Editor.next(); in particular, trying to compare the text in a verseNumber to the desired number (string.) This is regardless of whether Node.string() is used on the verseNumber, or if the "text" field of the verseNumber's text child is used for the comparison instead.