Don't change user selection on click of verse number
Since the verse transforms (join, add, etc.) no longer user the editor's selection and instead use a given path, the selection does not have to be on the verse number. When a verse number is clicked, focus is given to the menu that appears under the verse number, so the cursor disappears and the selection is irrelevant.