Focus the editor on versemenu hide()
If another component besides the editor was previously focused when a verse number is clicked (such as the demo input dropdown menu), the focus will return to that component when the verse menu is closed. This commit brings the focus back to the editor, albeit after a momentary (but visible) shift in focus back to the previously focused component.