Change \v and \rq StyleType from 'Character' to 'Paragraph' in usfm.sty
It appears from the usfm documentation that both \v and \rq start on a new line,
except when the \rq is a cross-reference content element. If the \rq
node is a cross-reference content element, it will be entirely contained
within the "content" field of the "x" element (as parsed by usfm-js),
thus the designation as a 'Paragraph' StyleType would have no effect.
Otherwise, if it is not a cross-reference content element, it is
convenient that \rq be labeled as a 'Paragraph' StyleType, since for our purposes,
a StyleType of 'Paragraph' indicates that this marker
should have its own line in the usfm.