Don't show "Remove" verse button if it is the only verse
Without doing this check, the code throws an error if the remove button is clicked because it tries to concatenate the current verse with the previous verse's contents (and there is no previous verse.)

Or, instead of not showing the "Remove" verse button, we could just remove the verse and NOT attempt to concatenate with the previous verse if there isn't one. Then there would no longer be even a single verse in the chapter.